Marie a plus d'erreurs, et plus d'id

+++ ProjetDassaultSystem-ejb/src/java/Session/UtilisateurSessionLocal.java
@@ -13,6 +13,6 @@ import javax.ejb.Local;
 @Local
 public interface UtilisateurSessionLocal {
 
-    int authentification(String login, String mdp);
+    long authentification(String login, String mdp);
     
 }
